backend_analysis/requirements.txt
# Các dependencies khác...
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.25
psycopg2-binary==2.9.10
requests==2.28.2
beautifulsoup4==4.12.0
apscheduler==3.10.1
httpx==0.23.3
python-dotenv==1.0.0
pydantic>=2.0.0
newspaper3k==0.2.8
lxml_html_clean==0.4.0
websockets==10.3
openpyxl>=3.0.0